Rewrite mobile friendliness check
